00{3,7}: use long instead of long long

+++ 0/0/7/007.c
@@ -7,7 +7,7 @@
 #define TARGET 10001
 
 int main() {
-  long long i, j;
+  long i, j;
   int n = 0;
   char *sieve;
